Split disk commands into modules
